Added images of completed recipe
Create the img file and download images of the completed recipe.
Add img element with src attribute with relative links to the images
and alt attribute to describe the image. Resize the cookie image size
using the width and height attributes after using an aspect ratio
re-sizing calculator for values.